pyproject.toml
[project]
name = "openclaw"
version = "1.0.0"
description = "Industrial AI gateway — intent-aware LLM routing for factory floor diagnostics"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "FactoryLM", email = "mike@factorylm.ai"}]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "groq>=0.4.0",
    "anthropic>=0.18.0",
    "openai>=1.10.0",
    "google-generativeai>=0.4.0",
    "python-telegram-bot>=20.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
plc = ["pymodbus>=3.6.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23.0", "ruff>=0.3.0"]

[project.scripts]
openclaw = "openclaw.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
